+++ src/libs/mzf/mzf.h
@@ -69,10 +69,10 @@ extern "C" {
 
     extern int mzf_read_header ( st_HANDLER *h, st_MZF_HEADER *mzfhdr );
     extern int mzf_write_header ( st_HANDLER *h, st_MZF_HEADER *mzfhdr );
-    extern int mzf_read_body_on_offset ( st_HANDLER *h, uint32_t offset, uint8_t *buffer, uint16_t buffer_size );
-    extern int mzf_read_body ( st_HANDLER *h, uint8_t *buffer, uint16_t buffer_size );
-    extern int mzf_write_body_on_offset ( st_HANDLER *h, uint32_t offset, uint8_t *buffer, uint16_t buffer_size );
-    extern int mzf_write_body ( st_HANDLER *h, uint8_t *buffer, uint16_t buffer_size );
+    extern int mzf_read_body_on_offset ( st_HANDLER *h, uint32_t offset, uint8_t *buffer, uint32_t buffer_size );
+    extern int mzf_read_body ( st_HANDLER *h, uint8_t *buffer, uint32_t buffer_size );
+    extern int mzf_write_body_on_offset ( st_HANDLER *h, uint32_t offset, uint8_t *buffer, uint32_t buffer_size );
+    extern int mzf_write_body ( st_HANDLER *h, uint8_t *buffer, uint32_t buffer_size );
     extern const char* mzf_error_message ( st_HANDLER *h, st_DRIVER *d );
 
 #define MZF_UINT8_FNAME(n) (uint8_t*)&n
